limit string_to_interpolation_type scope to fix aspect warnings.
